Add tests for spatial invariants
There are some useful generic spatial invariants for one dimensional
molecules that are very easy to compute. Testing these invariants should
easily reveal a number of easy mistakes in the various numerical
routines.

This change adds tests for translation and for inversion through the
molecule's midpoint. The parametrisation is very limited right now, but
we can add more later.