Refactor the tests against published results
I'm basically starting from nothing on the test coverage front. I think
locking down the final output of the program is going to be the most
efficient way to begin addressing that. I previously added some tests
against the published results to do this, but they were a bit ugly and a
little thin.

These tests are kind of difficult. The published numbers are basis set
converged results, and they were obtained from large and expensive
calculations. I've built some machinery to try and evaluate how much
we've converged when we use a small test basis set. Then we can compare
just what has converged, and if we don't have any converged digits we
just skip the test.

Doing this takes a long time though, so these tests are now skipped
unless an extra flag is passed in. It was also a giant pain to get
working for some reason, counting converged digits in code is really
annoying.